Moved string sequence functions to separate file
Changelog: None
Ticket:	ENT-4885
