+++ libutils/Makefile.am
@@ -74,6 +74,7 @@ libutils_la_SOURCES = \
 	regex.c regex.h \
 	ring_buffer.c ring_buffer.h \
 	sequence.c sequence.h \
+	string_sequence.c string_sequence.h \
 	set.c set.h \
 	stack.c stack.h \
 	statistics.c statistics.h \
