+++ libutils/sequence.h
@@ -226,57 +226,8 @@ void SeqClear(Seq *seq);
   */
 Seq *SeqGetRange(const Seq *seq, size_t start, size_t end);
 
-
 void SeqRemoveNulls(Seq *s);
 
-//////////////////////////////////////////////////////////////////////////////
-// SeqString - Sequence of strings (char *)
-//////////////////////////////////////////////////////////////////////////////
-
-/**
-  @brief Create a new Sequence from splitting a string on a fixed delimiter
-  @param [in] str String to split.
-  @param [in] delimiter The delimiter, a fixed string.
-  @return A pointer to the always created Sequence
-  */
-Seq *SeqStringFromString(const char *str, char delimiter);
-
-
-/**
- * @brief Return the total string length of a sequence of strings
- */
-int SeqStringLength(Seq *seq);
-
-
-/**
- * @brief Serializes a sequence of strings to a length prefixed format
- *
- * (De)Serialize uses a length prefixed format.
- * For every element in a string sequence,
- * the serialized output includes:
- * 1. 10 bytes of length prefix, where index 9 must be a space
- * 2. The data, with no escaping / modifications
- * 3. A single newline (\n) for readability
- * It is assumed that the sequence contains ascii printable
- * NUL terminated characters.
- */
-char *SeqStringSerialize(Seq *seq);
-
-/**
- * @brief Create a sequence of strings from the serialized format
- *
- * @param[in] serialized The input string, contents are copied
- * @return A sequence of new allocated strings
- */
-Seq *SeqStringDeserialize(const char *serialized);
-
-/**
- * @brief Convert the argc + argv in main function to a sequence of strings
- *
- * Does not copy, edit or free the elements in argv.
- * Shallow copy. SeqDestroy is safe, it will not free the contents of argv.
- *
- */
 Seq *SeqFromArgv(int argc, const char *const *argv);
 
 #endif
