+++ tests/unit/sequence_test.c
@@ -1,6 +1,7 @@
 #include <test.h>
 
 #include <sequence.c>
+#include <string_sequence.c>
 #include <alloc.h>
 
 static Seq *SequenceCreateRange(size_t initialCapacity, size_t start, size_t end)
